refactor

+++ tests/fixtures/test-setup.js
@@ -1,7 +1,6 @@
 const supertest = require('supertest')
 const path = require('path')
 const mongoose = require('mongoose')
-mongoose.Promise = global.Promise
 
 const db = require('../../utils/db')
 
@@ -11,30 +10,29 @@ beforeAll(async () => {
     global.seed_data = require('./seed_data')
     global.clone = require('rfdc')()
 
+    // mongo db - different database for each test suite
+    await db.connect(process.env.mongo_uri_test, {
+        dbName: `db-${path.basename(process.env.TEST_SUITE).split('.').join('-')}`
+    })
+
     // express server
     const server = await require(`${process.cwd()}/app`).create()
     global.request = supertest(server)
-
-    // mongo db - different database for each test suite
-    if (mongoose.connection.readyState === 0) {
-        await db.connect(process.env.mongo_uri_test, {
-            dbName: `db-${path.basename(process.env.TEST_SUITE).split('.').join('-')}`
-        })
-    }
 })
 
 beforeEach(async () => {
-    // clear db
-    const collections = Object.keys(mongoose.connection.collections)
-    for (let name of collections) {
-        let collection = mongoose.connection.collections[name]
-        await collection.deleteMany()
-    }
+    if (process.env.TEST_SUITE.includes('routes')) {
+        // clear db
+        const collections = Object.keys(mongoose.connection.collections)
+        for (let name of collections) {
+            let collection = mongoose.connection.collections[name]
+            await collection.deleteMany()
+        }
 
-    // populate seed data
-    await global.request.post('/api/books')
-        .set('Accept', 'application/json')
-        .send(global.seed_data.books)
+        // populate seed data
+        const books = require('../../models/books')
+        await books.create(global.seed_data.books)
+    }
 })
 
 afterAll(async () => {
